requirements.txt
boto3
pandas
numpy
scipy
matplotlib
notebook
dvc
python-box == 6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
scikit-learn
xgboost==1.6.2
fastapi==0.78.0
imblearn==0.0
python-dotenv==0.21.0
uvicorn==0.18.3
-e .